Fixed missle crash bug
I know there was a circumstance where the missiles were not getting dealt with correctly and causing the MUD to crash. I don't remember the specifics but I went through and made sure the missile is being extracted where ever it was being destroyed. Also had to reformat the file to make it easier to diff with other versions of SWR that had this bug fixed. Also removed some function prototypes that were never used.